requirement.txt
celery==5.0.5
decorator==4.4.2
Django==3.1.5
django_celery_beat
django-cors-headers==3.4.0
django-mssql-backend==2.8.
django-redis==4.12.1
django-redis-cache==3.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
python-dotenv==0.16.0
gunicorn==20.0.4
ipykernel==4.6.1
ipython==5.5.0
ipython-genutils==0.2.0
minio==7.0.3
psycopg2-binary
redis==3.5.3
redis-structures==0.1.7
requests==2.21.0
sqlparse==0.3.1
wfastcgi==3.0.0
